Shall add Test Division Expected/Unexpected

+++ src/main/java/ch/bbw/Calculator.java
@@ -15,8 +15,9 @@ public class Calculator {
     public int subtraktion(int val1, int val2) {
         return val1 - val2;
     }
-    public double division(double dividend, double divisor){
-        return dividend/divisor;
+
+    public int division(int dividend, int divisor) {
+        return dividend / divisor;
     }
 
 }
